chart_backend/requirements.txt
Django==4.2.2                   # Django framework
djangorestframework==3.14.0      # Django REST framework
django-cors-headers
